server/debug_db.py
import mysql.connector

DB_CONFIG = {
    'host': 'localhost',
    'user': 'root',
    'password': '',
    'database': 'landsnot_db',
}

try:
    conn = mysql.connector.connect(**DB_CONFIG)
    cursor = conn.cursor(dictionary=True)
    
    cursor.execute("DESCRIBE users;")
    print("USERS TABLE:")
    for row in cursor.fetchall():
        print(row)
        
    cursor.execute("DESCRIBE static_nodes;")
    print("\nSTATIC_NODES TABLE:")
    for row in cursor.fetchall():
        print(row)
        
except Exception as e:
    print(f"Error: {e}")
finally:
    if 'conn' in locals() and conn.is_connected():
        cursor.close()
        conn.close()
