server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
mysql-connector-python==9.0.0
joblib==1.4.2
pandas==2.2.2
numpy==1.26.4
httpx==0.27.0
pydantic==2.9.0
scikit-learn==1.5.1
bcrypt==4.2.0
PyJWT==2.9.0
